Mod: drag and drop listener now on TimedElementListView
The approach of putting the listener on each item view in the
ListAdapter was not satisfying. When the ListView recycles
item views, it works well, but when it creates a new View object
whereas the drag&drop already started, these new objects are not
elligible to receive drag&drop events.