Fix: drag and drop start listener was using a bad position value
As the listener was created only at View creation, and not View
recycling, it was using bad position and view values. Now the
listener is created each time a View is asked, even if it is recycled.
A better fix could be to not use an anonymous class so we could pass
the View as a parameter: the listener would be therefore set only
at View creation.